Use useRef instead of useState in the form component
