+++ src/components/ThoughtForm.jsx
@@ -1,18 +1,34 @@
-import PropTypes from "prop-types";
+import { useRef } from "react";
+
+import { useThoughtStore } from "../stores/useThoughtStore";
 import styles from "./ThoughtForm.module.css";
 
-const ThoughtForm = ({ input, onSubmit, onChange }) => {
-  const charCount = input.trim().length;
+const ThoughtForm = () => {
+  const { postThought, setError } = useThoughtStore();
+  const inputRef = useRef("");
+  const charCount = inputRef.current.value?.trim().length;
+
+  // Post request on form submit, also validate the input
+  const createThought = event => {
+    event.preventDefault();
+    const message = inputRef.current.value;
+    if (message.trim().length >= 5 && message.trim().length <= 140) {
+      postThought(message);
+      inputRef.current.value = "";
+    } else {
+      setError("Input invalid: You must type within 5 to 140 words");
+    }
+  };
   return (
-    <form className={styles.form} onSubmit={onSubmit}>
+    <form className={styles.form} onSubmit={createThought}>
       <p className={styles.title}>What&apos;s making you happy right now?</p>
       <label htmlFor="message"></label>
       <textarea
         className={styles.input}
         id="message"
         name="message"
-        value={input}
-        onChange={onChange}
+        ref={inputRef}
+        onChange={() => setError()}
         placeholder="Type your happy thought here!"
       />
       <p
@@ -33,9 +49,3 @@ const ThoughtForm = ({ input, onSubmit, onChange }) => {
 };
 
 export default ThoughtForm;
-
-ThoughtForm.propTypes = {
-  input: PropTypes.string.isRequired,
-  onSubmit: PropTypes.func.isRequired,
-  onChange: PropTypes.func.isRequired,
-};
