+++ src/components/ThoughtsCollection.jsx
@@ -1,5 +1,5 @@
 import Lottie from "lottie-react";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 
 import emptySearch from "../assets/emptySearch.json";
 import loadingAnimation from "../assets/loading_animation.json";
@@ -16,45 +16,20 @@ const ThoughtsCollection = () => {
     loading,
     error,
     fetchThoughts,
-    postThought,
-    setError,
     filter,
     likedThoughts,
     sentThoughts,
   } = useThoughtStore();
-  const [message, setMessage] = useState("");
-
-  //reset the error and record the value of text input
-  const handleInputChange = event => {
-    setError();
-    const userInput = event.target.value;
-    setMessage(userInput);
-  };
 
   //useEffect to render API data on mount
   useEffect(() => {
     fetchThoughts();
   }, [fetchThoughts]);
 
-  // Post request on form submit, also validate the input
-  const createThought = event => {
-    event.preventDefault();
-    if (message.trim().length >= 5 && message.trim().length <= 140) {
-      postThought(message);
-      setMessage("");
-    } else {
-      setError("Input invalid: You must type within 5 to 140 words");
-    }
-  };
-
   return (
     <div>
       <HandleError error={error} />
-      <ThoughtForm
-        onSubmit={createThought}
-        input={message}
-        onChange={handleInputChange}
-      />
+      <ThoughtForm />
       {thoughts && <Counter />}
       <div className={styles.thoughts}>
         {
@@ -75,10 +50,7 @@ const ThoughtsCollection = () => {
               ? sentThoughts
               : thoughts
             ).map(thought => (
-              <ThoughtCard
-                thought={thought}
-                key={thought._id}
-              />
+              <ThoughtCard thought={thought} key={thought._id} />
             ))
         }
         {loading && <Lottie animationData={loadingAnimation} loop={true} />}
